Removed start.js logic and moved it into index.js There was no need for a start.js - decided just to call it index.js
